+++ app/src/features/projects/components/ProjectFundingItemForm.test.tsx
@@ -1,5 +1,5 @@
 import { render, waitFor } from '@testing-library/react';
-import { IAutocompleteFieldOptionWithType } from 'components/fields/AutocompleteFieldWithType';
+import { IAutocompleteFieldOptionWithType } from 'features/projects/components/FundingSourceAutocomplete';
 import { Formik } from 'formik';
 import React from 'react';
 import { IInvestmentActionCategoryOption } from './ProjectFundingForm';
